refactor(den): remove duplicate den-v2 service
Keep services/den as the only hosted-worker app so Den no longer ships two app surfaces to maintain, and fix the Docker entrypoints needed to verify the canonical service.
