--- .github/workflows/deploy-den-v2.yml
@@ -1,19 +1,0 @@
-# Deploy Den v2 workflow intentionally disabled.
-#
-# Render is responsible for building and deploying the Daytona-first Den v2
-# control plane, so this repository workflow should not trigger deploys or
-# mutate Render environment configuration.
-
-name: Deploy Den v2 (disabled)
-
-on:
-  workflow_dispatch:
-
-jobs:
-  disabled:
-    if: false
-    runs-on: ubuntu-latest
-    steps:
-      - name: Disabled
-        run: |
-          echo "deploy-den-v2.yml is intentionally disabled; Render handles Den v2 deploys"
