--- services/den-v2/.env.example
@@ -1,73 +1,0 @@
-DATABASE_URL=
-DATABASE_HOST=
-DATABASE_USERNAME=
-DATABASE_PASSWORD=
-DB_MODE=
-BETTER_AUTH_SECRET=
-BETTER_AUTH_URL=http://localhost:8788
-DEN_BETTER_AUTH_TRUSTED_ORIGINS=http://localhost:3005,http://localhost:5173
-GITHUB_CLIENT_ID=
-GITHUB_CLIENT_SECRET=
-GOOGLE_CLIENT_ID=
-GOOGLE_CLIENT_SECRET=
-PORT=8788
-WORKER_PROXY_PORT=8789
-CORS_ORIGINS=http://localhost:3005,http://localhost:5173
-PROVISIONER_MODE=stub
-OPENWORK_DAYTONA_ENV_PATH=
-WORKER_URL_TEMPLATE=https://workers.example.com/{workerId}
-RENDER_API_BASE=https://api.render.com/v1
-RENDER_API_KEY=
-RENDER_OWNER_ID=
-RENDER_WORKER_REPO=https://github.com/different-ai/openwork
-RENDER_WORKER_BRANCH=dev
-RENDER_WORKER_ROOT_DIR=services/den-worker-runtime
-RENDER_WORKER_PLAN=standard
-RENDER_WORKER_REGION=oregon
-RENDER_WORKER_OPENWORK_VERSION=0.11.113
-RENDER_WORKER_NAME_PREFIX=den-worker
-RENDER_WORKER_PUBLIC_DOMAIN_SUFFIX=openwork.studio
-RENDER_CUSTOM_DOMAIN_READY_TIMEOUT_MS=240000
-RENDER_PROVISION_TIMEOUT_MS=900000
-RENDER_HEALTHCHECK_TIMEOUT_MS=180000
-RENDER_POLL_INTERVAL_MS=5000
-VERCEL_API_BASE=https://api.vercel.com
-VERCEL_TOKEN=
-VERCEL_TEAM_ID=
-VERCEL_TEAM_SLUG=prologe
-VERCEL_DNS_DOMAIN=openwork.studio
-POLAR_FEATURE_GATE_ENABLED=false
-POLAR_API_BASE=https://api.polar.sh
-POLAR_ACCESS_TOKEN=
-POLAR_PRODUCT_ID=
-POLAR_BENEFIT_ID=
-POLAR_SUCCESS_URL=http://localhost:8788
-POLAR_RETURN_URL=http://localhost:8788
-DAYTONA_API_URL=https://app.daytona.io/api
-DAYTONA_API_KEY=
-DAYTONA_TARGET=
-DAYTONA_SNAPSHOT=
-DAYTONA_SANDBOX_IMAGE=node:20-bookworm
-DAYTONA_SANDBOX_CPU=2
-DAYTONA_SANDBOX_MEMORY=4
-DAYTONA_SANDBOX_DISK=8
-DAYTONA_SANDBOX_PUBLIC=false
-DAYTONA_SANDBOX_AUTO_STOP_INTERVAL=0
-DAYTONA_SANDBOX_AUTO_ARCHIVE_INTERVAL=10080
-DAYTONA_SANDBOX_AUTO_DELETE_INTERVAL=-1
-DAYTONA_SIGNED_PREVIEW_EXPIRES_SECONDS=86400
-DAYTONA_WORKER_PROXY_BASE_URL=https://workers.den.openworklabs
-DAYTONA_SANDBOX_NAME_PREFIX=den-daytona-worker
-DAYTONA_VOLUME_NAME_PREFIX=den-daytona-worker
-DAYTONA_WORKSPACE_MOUNT_PATH=/workspace
-DAYTONA_DATA_MOUNT_PATH=/persist/openwork
-DAYTONA_RUNTIME_WORKSPACE_PATH=/tmp/openwork-workspace
-DAYTONA_RUNTIME_DATA_PATH=/tmp/openwork-data
-DAYTONA_SIDECAR_DIR=/tmp/openwork-sidecars
-DAYTONA_OPENWORK_PORT=8787
-DAYTONA_OPENCODE_PORT=4096
-DAYTONA_OPENWORK_VERSION=
-DAYTONA_CREATE_TIMEOUT_SECONDS=300
-DAYTONA_DELETE_TIMEOUT_SECONDS=120
-DAYTONA_HEALTHCHECK_TIMEOUT_MS=300000
-DAYTONA_POLL_INTERVAL_MS=5000
