--- services/den-v2/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@openwork/den-v2",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "npm run build:den-db && OPENWORK_DEV_MODE=1 tsx watch src/index.ts",
-    "build": "npm run build:den-db && tsc -p tsconfig.json",
-    "build:den-db": "npm --prefix ../../packages/den-db run build",
-    "start": "node dist/index.js",
-    "db:migrate:sql": "node scripts/run-sql-migrations.mjs",
-    "test:smoke:daytona": "pnpm build && node scripts/daytona-provisioner-smoke.mjs",
-    "test:e2e:daytona": "node scripts/e2e-daytona-worker.mjs",
-    "test:e2e:worker-limit": "node scripts/e2e-worker-limit.mjs",
-    "db:generate": "drizzle-kit generate",
-    "db:migrate": "drizzle-kit migrate",
-    "auth:generate": "npx @better-auth/cli@latest generate --config src/auth.ts --output src/db/better-auth.schema.ts --yes"
-  },
-  "dependencies": {
-    "@daytonaio/sdk": "^0.150.0",
-    "better-auth": "^1.4.18",
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
-    "drizzle-orm": "^0.45.1",
-    "express": "^4.19.2",
-    "mysql2": "^3.11.3",
-    "zod": "^4.3.6"
-  },
-  "devDependencies": {
-    "@types/cors": "^2.8.17",
-    "@types/express": "^4.17.21",
-    "@types/node": "^20.11.30",
-    "drizzle-kit": "^0.31.9",
-    "tsx": "^4.15.7",
-    "typescript": "^5.5.4"
-  }
-}
