--- services/den-v2/src/db/drizzle.ts
@@ -1,1 +1,0 @@
-export { and, asc, desc, eq, gt, isNotNull, isNull, sql } from "../../../../packages/den-db/dist/drizzle.js"
