--- services/den-v2/src/db/index.ts
@@ -1,9 +1,0 @@
-import { createDenDb, isTransientDbConnectionError } from "../../../../packages/den-db/dist/index.js"
-import { env } from "../env.js"
-
-export const { db } = createDenDb({
-  databaseUrl: env.databaseUrl,
-  mode: env.dbMode,
-  planetscale: env.planetscale,
-})
-export { isTransientDbConnectionError }
