--- services/den-v2/src/db/schema.ts
@@ -1,1 +1,0 @@
-export * from "../../../../packages/den-db/dist/schema.js"
