--- services/den-v2/src/db/typeid.ts
@@ -1,1 +1,0 @@
-export * from "../../../../packages/utils/dist/typeid.js"
